alcohol_detection_system/requirements.txt
alembic==1.14.0
atomicwrites==1.4.1
attrs==24.2.0
blinker==1.9.0
click==8.1.7
colorama==0.4.6
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
marshmallow==3.14.1
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.9
py==1.11.0
PyJWT==2.9.0
pytest==6.2.5
pytest-flask==1.2.0
python-dotenv==0.19.0
six==1.16.0
SQLAlchemy==1.4.23
toml==0.10.2
typing_extensions==4.12.2
Werkzeug==2.0.1
